refactor: settings that configure something, on a page shaped like the work
Three of the seven things the operator flagged, and one correction to my
own reading of them.

**Two settings had no reader anywhere.** `default_container` named a v1
image (`vllm-node`) that engines replaced; `default_gpu_mem_util` was
superseded by the recipe's own `gpu_memory_utilization` and by the deploy
form. Neither ever reached a launch. This is the same defect
`docker.cluster_image`, `ray_port` and `gpu_count` were removed for — a
form that looks like configuration and configures nothing — and these two
were missed. Their tests now assert the absence and say why, so they do not
come back.

**`cluster_enabled` is not one of them, and I said it was.** A grep I cut
short showed no reader and I reported it as inert. `RecipesPage.tsx:227`
reads it to decide whether a `cluster_only` recipe is offered or listed as
unavailable. What is actually wrong is narrower: the value ships in two
payloads and only one is read — `/api/settings` feeds the recipes page,
while `/api/config` carried a second copy that `lib/config.tsx` types and
nothing consults. Two publishers for one switch is how a live setting comes
to look dead. The switch stays, the duplicate goes.

**The Cluster tab is gone.** Its switch was a feature switch, so it joins
the other feature switches; the rest of the tab described the machines, and
the machines are managed on `/cluster`. Network discovery moved there as
`components/NetworkDiscovery.tsx`, next to the node registry — a reading
about this host beside the list of hosts, rather than two pages away under
a heading that also held ports and shm sizes. Its tests moved with it.

**Benchmarking was filed under "Timeouts and History".** It is not a
timeout: `Layout.tsx` and `App.tsx` read it to decide whether the
Benchmarking route and sidebar entry exist at all. The new Features tab is
for exactly that kind of switch — the ones that change the shape of the app
rather than the behaviour of a deployment — and MCP is reported there
beside it, since it answers the same question about whether a way in
exists.

**One card per row.** Every tab was `lg:grid-cols-2`, which existed so that
tabs holding two small cards did not look empty. A settings form is read one
field at a time and a second column doubles the eye travel for every field.
Environment keeps two columns: it is a reference table, not a form.
